Scripts/requirements.txt
pywinauto==0.6.8
Pillow==9.1.0
colorama==0.4.4
mss==6.1.0
pygetwindow==0.0.9
